% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoAnalysis_LIB.R
\name{dataframeFromJSON}
\alias{dataframeFromJSON}
\title{helper function fto parse the json results to data frames}
\usage{
dataframeFromJSON(l)
}
\arguments{
\item{json}{text}
}
\value{
data frame from json
}
\description{
Microsoft  API returns (well-)structured JSON, this function parses it into data frames
}

