% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxford_extension.R
\name{getDescriptionResponse}
\alias{getDescriptionResponse}
\title{image description}
\usage{
getDescriptionResponse(img.path, key, maxCandidates = 3)
}
\arguments{
\item{key}{for the vision api}

\item{url}{to image}
}
\value{
data frame with image description
}
\description{
provide path to image, get a description of the image back
}
\examples{
getDescriptionResponse('out/snap00169.png', visionKey, maxCandidates)

}

