% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxford_extension.R
\name{getDescriptionResponseURL}
\alias{getDescriptionResponseURL}
\title{image description}
\usage{
getDescriptionResponseURL(img.url, key, maxCandidates = 3)
}
\arguments{
\item{key}{for the vision api}

\item{url}{to image}
}
\value{
data frame with image description
}
\description{
provide url to image, get a description of the image back
}
\examples{
getDescriptionResponseURL("http://sizlingpeople.com/wp-content/uploads/2015/10/Kim-Kardashian-2015-21.jpg", visionKey)

}

