% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxford_extension.R
\name{getDomainModelResponse}
\alias{getDomainModelResponse}
\title{get a model based response back}
\usage{
getDomainModelResponse(img.path, key, model = "celebrities")
}
\arguments{
\item{key}{for the vision api}

\item{model}{see getDomainModels()}

\item{image}{path}
}
\value{
data frame model specific features
}
\description{
provide a specific model (e.g. 'celebrities') and get the classification back
}
\examples{
getDomainModelResponseURL("out/image.png", visionKey, model)

}

