% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxford_extension.R
\name{getDomainModelResponseURL}
\alias{getDomainModelResponseURL}
\title{get a model based response back}
\usage{
getDomainModelResponseURL(img.url, key, model = "celebrities")
}
\arguments{
\item{key}{for the vision api}

\item{model}{see getDomainModels()}
}
\value{
data frame model specific features
}
\description{
provide a specific model (e.g. 'celebrities') and get the classification back
}
\examples{
getDomainModelResponseURL(img.url, visionKey, model)

}

