% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxford_extension.R
\name{getDomainModels}
\alias{getDomainModels}
\title{get currently available domain specific models!}
\usage{
getDomainModels(key)
}
\arguments{
\item{key}{for the vision api}
}
\value{
data frame with domain specific models
}
\description{
provide visionkey to get available models
}
\examples{
getDomainModels(visionKey)

}

