% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoAnalysis_LIB.R
\name{getEmotionResponseURL}
\alias{getEmotionResponseURL}
\title{emotion detection in images with a human faces}
\usage{
getEmotionResponseURL(img.url, key)
}
\arguments{
\item{key}{for the emotion api}

\item{url}{to image}
}
\value{
data frame with emotion scores
}
\description{
provide an url to an image, get emotion scores back for each face.
}
\examples{
getEmotionResponseURL("http://sizlingpeople.com/wp-content/uploads/2015/10/Kim-Kardashian-2015-21.jpg", emotionKey)

}

