% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoAnalysis_LIB.R
\name{getFaceResponseURL}
\alias{getFaceResponseURL}
\title{get face attributes, age, gender, faceid}
\usage{
getFaceResponseURL(img.url, key)
}
\arguments{
\item{key}{for the face api}

\item{url}{to image}
}
\value{
data frame with face attributes, age, gender, faceid
}
\description{
url to image, get text back!
}
\examples{
getFaceResponseURL("http://sizlingpeople.com/wp-content/uploads/2015/10/Kim-Kardashian-2015-21.jpg", facekey)

}

