% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoAnalysis_LIB.R
\name{getOCRResponse}
\alias{getOCRResponse}
\title{OCR recognition function}
\usage{
getOCRResponse(img.path, visionKey, language = "de")
}
\arguments{
\item{language}{settings default is DE}

\item{path}{to local image}

\item{key}{to vision api}
}
\value{
data frame of text blocks
}
\description{
upload image, get text back!
}
\examples{
getOCRResponse("out/snap00169.png", visionKey)
}

