% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxford_extension.R
\name{getTaggingResponse}
\alias{getTaggingResponse}
\title{tags images}
\usage{
getTaggingResponse(img.path, key)
}
\arguments{
\item{key}{for the vision api}
}
\value{
data frame with image description
}
\description{
provide  image, get a tags/labels back
}
\examples{
getTaggingResponseURL("out/image.png", visionKey)

}

