% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxford_extension.R
\name{getTaggingResponseURL}
\alias{getTaggingResponseURL}
\title{image tags}
\usage{
getTaggingResponseURL(img.url, key)
}
\arguments{
\item{key}{for the vision api}

\item{url}{to image}
}
\value{
data frame with image description
}
\description{
provide url to image, get a tags/labels back
}
\examples{
getTaggingResponseURL("http://sizlingpeople.com/wp-content/uploads/2015/10/Kim-Kardashian-2015-21.jpg", visionKey)

}

