% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoAnalysis_LIB.R
\name{getVideoMotion}
\alias{getVideoMotion}
\title{main call to video motion API}
\usage{
getVideoMotion(video.path, key)
}
\arguments{
\item{key}{for the video api}

\item{path}{to local video}
}
\value{
data frame with video motion results
}
\description{
might take a while
}
\examples{
getVideoMotion("video.mp4", videoKey)

}

