% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoAnalysis_LIB.R
\name{getVideoResponse}
\alias{getVideoResponse}
\title{main call to video API}
\usage{
getVideoResponse(video.path, key)
}
\arguments{
\item{key}{for the video api}

\item{path}{to local video}
}
\value{
data frame with video results
}
\description{
might take a while
}
\examples{
getVideoResponse("video.mp4", videoKey)

}

