% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoAnalysis_LIB.R
\name{getVisionResponseURL}
\alias{getVisionResponseURL}
\title{image recognition and object identification}
\usage{
getVisionResponseURL(img.url, key)
}
\arguments{
\item{key}{for the vision api}

\item{url}{to image}
}
\value{
data frame with image attributes
}
\description{
provide url to image, a description of the image back
}
\examples{
getVisionResponseURL("http://sizlingpeople.com/wp-content/uploads/2015/10/Kim-Kardashian-2015-21.jpg", facekey)

}

