% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoAnalysis_LIB.R
\name{getVisionResponse}
\alias{getVisionResponse}
\title{image recognition and object identification}
\usage{
getVisionResponse(img.path, key)
}
\arguments{
\item{key}{for the vision api}

\item{path}{to local image}
}
\value{
data frame with image attributes
}
\description{
upload image, a description of the image back
}
\examples{
getVisionResponse("out/snap00169.png", facekey)

}

